requirements.txt
opencv-python
numpy
torch
transformers
langchain
pinecone-client
sentence-transformers
python-dotenv